feat: This is an automated commit from the student work uploader.
The content of this commit is the work of students. These changes was created during the System Development 2 course of the University of Szeged as a project assignment. The work was checked by the author (the tutor) of this commit.

The original authors of this changes are: Attila Szabolics <szabolics.attila@gmail.com>, Simon Soma <somas491@gmail.com>.